Refactors FeedRootViewModel to FeedMainViewModel
Renames `FeedRootViewModel` to `FeedMainViewModel` for clarity.

Updates `FeedMainView` to use the renamed ViewModel and initializes the `items` with a fetch limit of 10.

Improves the initial data loading logic by ensuring it only runs once.
Adds a check to prevent refreshing while a refresh is already in progress.

Streamlines navigation title.

Removes context menu sheet options for favoriting, reading posts, as well as the refreshable modifier and topBarLeading toolbarItem as the button functionality has been removed